`get_posts` uses Discussion.posts().
It uses `Discussion.posts()` for the upper-level of posts, the ones that will
be sorted correctly, and post.descendants for each of one of the returned
posts.

The descendants are included in the response only to satisfy the existing
front-end threading algorithms. The problem is that the descendants will not be
sorted properly, and it may be hard to find new messages. At least by using
`Discussion.posts()` for the upper-level, the posts with new descendants will
at least be at the top of the list.