Extract.id is an Integer again.
I was getting a vague transaction error when I tried to save an Extract to the
database. I reinstated the Integer id column and the error stopped happening.

The UUID that the front-end sends is not needed for anything in the back end
and is only meant for the backbone environment. The front-end code is copying
the [automatically generated](http://backbonejs.org/#Model-cid) client-side id
into the new model instance's id parameter, causing backbone to think it needs
to `PUT` when the instance's `save` method is called. This behaviour is wrong and
unnecessary.